refactor(consensus): consolidate AuthoritySet into own module
Move AuthoritySet from block_header.rs to authority_set.rs. Replace
SubsetAuthorities in cordial_knowledge.rs and raw [u64; 4] bitmask
helpers in network/mod.rs with the shared AuthoritySet type.
