+++ crates/starfish/core/src/network/mod.rs
@@ -32,6 +32,7 @@ use starfish_config::AuthorityIndex;
 
 use crate::{
     Round, VerifiedBlockHeader,
+    authority_set::AuthoritySet,
     block_header::{BlockRef, VerifiedBlock},
     commit::{CommitRange, TrustedCommit},
     error::{ConsensusError, ConsensusResult},
@@ -293,41 +294,16 @@ pub(crate) struct SerializedBlockBundleParts {
     pub(crate) serialized_block: Bytes,
     pub(crate) serialized_headers: Vec<Bytes>,
     pub(crate) serialized_shards: Vec<Bytes>,
-    pub(crate) useful_headers_authors_bitmask: [u64; 4],
-    pub(crate) useful_shards_authors_bitmask: [u64; 4],
-}
-
-fn authority_set_to_bitmask(authorities: &BTreeSet<AuthorityIndex>) -> [u64; 4] {
-    let mut bitmask = [0u64; 4];
-    for authority_index in authorities {
-        let index = authority_index.value();
-        let array_index = index / 64;
-        let bit_pos = index % 64;
-        bitmask[array_index] |= 1u64 << bit_pos;
-    }
-    bitmask
-}
-
-fn bitmask_to_authority_set(bitmask: [u64; 4]) -> BTreeSet<AuthorityIndex> {
-    let mut set = BTreeSet::new();
-    for (array_index, &bits) in bitmask.iter().enumerate() {
-        let mut bits = bits;
-        let base = array_index * 64;
-        while bits != 0 {
-            let bit = bits.trailing_zeros() as usize;
-            set.insert(AuthorityIndex::from((base + bit) as u8));
-            bits &= bits - 1;
-        }
-    }
-    set
+    pub(crate) useful_headers_authors_bitmask: AuthoritySet,
+    pub(crate) useful_shards_authors_bitmask: AuthoritySet,
 }
 
 impl SerializedBlockBundleParts {
     pub(crate) fn useful_headers_authors(&self) -> BTreeSet<AuthorityIndex> {
-        bitmask_to_authority_set(self.useful_headers_authors_bitmask)
+        self.useful_headers_authors_bitmask.to_btreeset()
     }
     pub(crate) fn useful_shards_authors(&self) -> BTreeSet<AuthorityIndex> {
-        bitmask_to_authority_set(self.useful_shards_authors_bitmask)
+        self.useful_shards_authors_bitmask.to_btreeset()
     }
 }
 
@@ -350,8 +326,8 @@ impl TryFrom<VerifiedBlock> for SerializedBlockBundleParts {
             serialized_block: Bytes::from(bytes),
             serialized_headers: vec![],
             serialized_shards: vec![],
-            useful_headers_authors_bitmask: [0u64; 4],
-            useful_shards_authors_bitmask: [0u64; 4],
+            useful_headers_authors_bitmask: AuthoritySet::new(),
+            useful_shards_authors_bitmask: AuthoritySet::new(),
         })
     }
 }
@@ -375,12 +351,10 @@ impl TryFrom<BlockBundle> for SerializedBlockBundleParts {
             serialized_block: Bytes::from(bytes),
             serialized_headers: serialized_block_headers,
             serialized_shards: block_bundle.serialized_shards,
-            useful_headers_authors_bitmask: authority_set_to_bitmask(
+            useful_headers_authors_bitmask: AuthoritySet::from(
                 &block_bundle.useful_headers_authors,
             ),
-            useful_shards_authors_bitmask: authority_set_to_bitmask(
-                &block_bundle.useful_shards_authors,
-            ),
+            useful_shards_authors_bitmask: AuthoritySet::from(&block_bundle.useful_shards_authors),
         })
     }
 }
